Tuning the number of packets for one read in StreamingSource
Read performance in StreamingSource became worse from LMR1 because
the design was changed to asynchronous from synchronous.
As a result, the generation of access unit does not make it
in time for some multiplexed contents, and this causes frame drop.
This patch improves read performance by increasing the number of
packets for one read.

Change-Id: I14955da5a9c5415e35905cc95f1471e0811a4991